+++ examples/seismic/__init__.py
@@ -1,4 +1,5 @@
 from .model import *  # noqa
 from .source import *  # noqa
 from .plotting import *  # noqa
+from .preset_models import *  # noqa
 from .utils import *  # noqa
