+++ examples/seismic/elastic/elastic_example.py
@@ -50,8 +50,8 @@ def run(shape=(50, 50), spacing=(20.0, 20.0), tn=1000.0,
 def test_elastic():
     _, _, _, [rec1, rec2, v, tau] = run()
     norm = lambda x: np.linalg.norm(x.data.reshape(-1))
-    assert np.isclose(norm(rec1), 29.8925, atol=1e-3, rtol=0)
-    assert np.isclose(norm(rec2), 1.20394, atol=1e-3, rtol=0)
+    assert np.isclose(norm(rec1), 29.6567, atol=1e-3, rtol=0)
+    assert np.isclose(norm(rec2), 1.19761, atol=1e-3, rtol=0)
 
 
 if __name__ == "__main__":
